Collect pipeline input to String assertion
Assertions should provide a unified way of consuming input. The
way currently used by Pester results in confusing results where
1,2,3 | Should Be 3 passes the test. To avoid that all the pipeline
input is consumed first and then asserted on. Another thing is
needed for this to work reliably and that is putting the actual
on the right side of the equality check to avoid filtering the array
instead of comparing it.

Remove broken test and writing