Remove int, decimal, and double assertions as they are too specific.
Originally I planned every primitive type to have it's set of assertions, but always thinking about types is impractical. There should be Assert-Range that will deal with both big ranges (for checking that we were somewhere betwwen 1 and 3 GB of memory), and small ranges (to make assertions on imprecise values possible). Possibly using Assert-Range for chars might be considered, but what about utf, will that behave predictably and is it even useful?