Fix comparison of collections with two or more equivalent items
The collection algorithm was not stopping when the item was found in the collection, so all items that matched the expected item were marked as taken, so the subsequent iterations found the item, but it was rejected because it was already marked as equivalent to another item. To fix this when an item is found in Actual, we immediately move on to the next item in Expected. I also changed the order of the condition in to reject item from comparison when it was already used without comparing it again. And improved the equivalency message to show which expected item was linked to which actual item.